Updated Auth Refresh mechanism using AuthSession (#282)
After the upgrade from expo-app-auth to expo-auth-session, refreshToken for Google auth is no longer easy to obtain. Now we have to switch our auth flow entirely since expo-auth-session returns null for idToken and refreshToken. The new mechanism involves requesting an authorization_code and then exchanging it to an idToken and refreshToken.

Co-authored-by: Ashay Parikh <ashayparikh22@gmail.com>